# Optimize `alter`