Fixes #392.