@sjakobi 2022-04-03
The benchmark results are largely good, but some are too noisy, particularly `HashMap.insert`.

`master`:

```
    alterInsert-dup
      String:     OK (2.54s)
        583  μs ± 4.6 μs
      ByteString: OK (2.41s)
        275  μs ± 1.9 μs
      Int:        OK (4.01s)
        470  μs ± 1.8 μs
    alterDelete
      String:     OK (3.56s)
        833  μs ± 3.1 μs
      ByteString: OK (2.92s)
        674  μs ± 3.5 μs
      Int:        OK (3.95s)
        463  μs ± 1.4 μs
    alterDelete-miss
      String:     OK (1.87s)
        211  μs ± 1.6 μs
      ByteString: OK (2.50s)
        141  μs ± 1.1 μs
      Int:        OK (2.29s)
        259  μs ± 1.7 μs
```

This branch (1059d3c):

```
    alterInsert-dup
      String:     OK (2.14s)
        491  μs ± 3.5 μs
      ByteString: OK (2.33s)
        265  μs ± 1.5 μs
      Int:        OK (6.29s)
        373  μs ± 1.1 μs
    alterDelete
      String:     OK (1.86s)
        838  μs ± 6.5 μs
      ByteString: OK (2.76s)
        636  μs ± 2.7 μs
      Int:        OK (2.02s)
        454  μs ± 2.8 μs
    alterDelete-miss
      String:     OK (2.00s)
        225  μs ± 1.5 μs
      ByteString: OK (2.04s)
        114  μs ± 780 ns
      Int:        OK (2.20s)
        248  μs ± 1.4 μs
```