@treeowl 2022-05-10
Why doesn't this just do one pass? We can use manual worker-wrapper to try to unbox the `Maybe`s going in and out of the passed function.